services/nats-publisher/requirements.txt
nats-py>=2.7.0
orjson>=3.9.0
